requirements.txt
streamlit
ultralytics
opencv-python
numpy
easyocr
matplotlib
pillow
plotly
pandas 